remove shakemap map
