+++ src/app/shakemap/intensity/intensity.component.scss
@@ -1,0 +1,4 @@
+shared-map {
+    display: block;
+    height: 500px;
+}
