--- src/app/shakemap/map/map.component.html
@@ -1,10 +1,0 @@
-<ng-container *ngIf="event && shakemap">
-  <a
-      [routerLink]="'../../map'"
-      [queryParams]="queryParams">
-    <shared-map
-        [overlays]="shakemap | shakemapOverlays:enableOverlay"
-        [showScaleControl]="true">
-    </shared-map>
-  </a>
-</ng-container>
