--- src/app/shakemap/map/map.component.scss
@@ -1,4 +1,0 @@
-shared-map {
-  display: block;
-  height: 500px;
-}
