+++ src/app/shared/map/map.component.spec.ts
@@ -81,7 +81,7 @@ describe('MapComponent', () => {
       expect(component.getOverlayBounds).not.toHaveBeenCalled();
     });
 
-    it('uses overlay bounds', (done) => {
+    it('uses overlay bounds', () => {
       const bounds = [1, 2, 3];
       spyOn(component, 'getOverlayBounds').and.returnValue(bounds);
       spyOn(component.map, 'fitBounds').and.returnValue(null);
@@ -92,11 +92,10 @@ describe('MapComponent', () => {
 
       expect(component.getOverlayBounds).toHaveBeenCalled();
       expect(component.map.fitBounds).toHaveBeenCalledWith(bounds);
-      done();
 
     });
 
-    it('defaults to world', (done) => {
+    it('defaults to world', () => {
       spyOn(component, 'getOverlayBounds').and.returnValue(null);
       spyOn(component.map, 'fitBounds').and.returnValue(null);
 
@@ -107,7 +106,6 @@ describe('MapComponent', () => {
       expect(component.getOverlayBounds).toHaveBeenCalled();
       expect(component.map.fitBounds.calls.first().args[0]).toEqual(
           [[85, 180], [-85, 180]]);
-      done();
     });
   });
